+++ src/main/java/net/brnbrd/delightful/common/item/food/CompatDrinkItem.java
@@ -2,14 +2,7 @@ package net.brnbrd.delightful.common.item.food;
 
 import net.brnbrd.delightful.common.item.ICompat;
 import net.brnbrd.delightful.compat.Modid;
-import net.minecraft.network.chat.Component;
 import net.minecraft.world.item.Item;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.TooltipFlag;
-import net.minecraft.world.level.Level;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-import java.util.List;
 
 public class CompatDrinkItem extends DrinkItem implements ICompat {
 	private final Modid[] modid;
@@ -22,12 +15,5 @@ public class CompatDrinkItem extends DrinkItem implements ICompat {
 	@Override
 	public Modid[] getModid() {
 		return modid;
-	}
-
-	@Override
-	public void appendHoverText(@NotNull ItemStack stack, @Nullable Level level, @NotNull List<Component> comps, @NotNull TooltipFlag pIsAdvanced) {
-		if (this.enabledText(comps)) {
-			super.appendHoverText(stack, level, comps, pIsAdvanced);
-		}
 	}
 }