+++ src/main/java/net/brnbrd/delightful/common/item/food/DrinkItem.java
@@ -1,6 +1,7 @@
 package net.brnbrd.delightful.common.item.food;
 
 import net.brnbrd.delightful.common.item.IConfigured;
+import net.minecraft.network.chat.Component;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResultHolder;
 import net.minecraft.world.entity.LivingEntity;
@@ -8,9 +9,12 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.ItemUtils;
+import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.level.Level;
-import org.jetbrains.annotations.NotNull;
 import vectorwing.farmersdelight.common.item.DrinkableItem;
+import java.util.List;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 public class DrinkItem extends DrinkableItem implements IConfigured {
 	public DrinkItem(Item.Properties properties, boolean hasPotionEffectTooltip, boolean hasCustomTooltip) {
@@ -31,5 +35,12 @@ public class DrinkItem extends DrinkableItem implements IConfigured {
 		float heal = this.getHeal();
 		if (heal > 0F) consumer.heal(heal);
 		super.affectConsumer(stack, worldIn, consumer);
+	}
+
+	@Override
+	public void appendHoverText(@NotNull ItemStack stack, @Nullable Level level, @NotNull List<Component> comps, @NotNull TooltipFlag pIsAdvanced) {
+		if (this.enabledText(comps)) {
+			super.appendHoverText(stack, level, comps, pIsAdvanced);
+		}
 	}
 }